pyproject.toml
[project]
    authors = [
        { name = "Brian Welman", email = "brianallisterwelman@gmail.com" },
        { name = "Casey Chuma", email = "caseymovex@gmail.com" },
    ]
    name = "astro"
    dynamic = ["version"]
    description = "Radio Astronomy AI Assistant"
    readme = "README.md"
    license = { text = "MIT" }
    requires-python = ">=3.11"
    keywords = ["ai", "astronomy", "chat", "llm", "agent", "assistant"]
    classifiers = [
        "Development Status :: 4 - Beta",
        "Intended Audience :: Science/Research",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Topic :: Scientific/Engineering :: Astronomy",
        "Topic :: Scientific/Engineering :: Artificial Intelligence",
        "Typing :: Typed",
    ]
    dependencies = [
        "anthropic>=0.62.0",
        "appdirs>=1.4.4",
        "colour>=0.1.5",
        "distro>=1.9.0",
        "docstring-parser>=0.16",
        "jinja2>=3.1.6",
        "ollama>=0.5.3",
        "openai>=1.75.0",
        "prompt-toolkit>=3.0.52",
        "pydantic-ai>=1.0.15",
        "pydantic-graph>=1.0.15",
        "pydantic[email]>=2.10.3",
        "pygments>=2.19.2",
        "python-dotenv>=1.0.1",
        "python-frontmatter>=1.1.0",
        "pyyaml>=6.0.2",
        "rich>=13.9.4",
        "typer>=0.19.2",
    ]

[project.urls]
    Homepage = "https://github.com/kwazzi-jack/astro"
    Repository = "https://github.com/kwazzi-jack/astro"
    Issues = "https://github.com/kwazzi-jack/astro/issues"


[build-system]
    requires      = ["hatchling", "hatch-vcs"]
    build-backend = "hatchling.build"

[tool.hatch.version]
    source = "vcs"

[tool.hatch.build.hooks.vcs]
    version-file = "astro/__version__.py"

[tool.hatch.build.targets.sdist]
    exclude = [
        "examples/",
        "tests/",
        "htmlcov/",
        "trash/",
        ".pytest_cache/",
        "*.pyc",
        "__pycache__/",
    ]

[tool.hatch.build.targets.wheel]
    packages = ["astro"]
    exclude = [
        "examples/",
        "tests/",
    ]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "mdformat>=0.7.22",
    "mdformat-gfm>=0.4.1",
    "mdformat-tables>=1.0.0",
    "pytest-ruff>=0.5",
    "pytest-black>=0.6.0",
    "pytest-mypy>=1.0.1",
    "freezegun>=1.5.5",
    "pillow>=12.0.0",
]

[project.scripts]
astro-cli = "astro._bootstrap:_run_astro_cli"
